Floatingpoint movement

+++ src/Hazard.h
@@ -76,9 +76,9 @@ public:
 
 	FPoint pos;
 	FPoint speed;
-	int base_speed;
+	float base_speed;
 	int lifespan; // ticks down to zero
-	int radius;
+	float radius;
 	int power_index;
 
 	int animationKind;	// direction or other, it is a specific value according to
